+++ ttnn/cpp/ttnn/operations/experimental/ccl/all_gather_async/all_gather_async.cpp
@@ -81,30 +81,5 @@ std::vector<ttnn::Tensor> ExecuteAllGatherAsync::invoke(
         num_preferred_links,
         subdevice_id);
 }
-// Add the implementation for ExecuteAllToAllAsync::invoke
-ttnn::Tensor ExecuteAllToAllAsync::invoke(
-    const ttnn::Tensor& input_tensor,
-    ttnn::Tensor& persistent_intermediate_buffer,
-    ttnn::Tensor& persistent_output_buffer,
-    const int32_t in_dim,
-    const int32_t out_dim,
-    const GlobalSemaphore& multi_device_global_semaphore,
-    const uint32_t num_links,
-    const std::optional<ttnn::MemoryConfig>& memory_config,
-    const ttnn::ccl::Topology topology,
-    std::optional<tt::tt_metal::SubDeviceId> subdevice_id) {
-    // Forward the call to the underlying all_to_all_async implementation (to be created)
-    return ttnn::operations::experimental::ccl::all_to_all_async(
-        input_tensor,
-        persistent_intermediate_buffer,
-        persistent_output_buffer,
-        in_dim,
-        out_dim,
-        multi_device_global_semaphore,
-        num_links,
-        memory_config,
-        topology,
-        subdevice_id);
-}
 
 }  // namespace ttnn::operations::experimental::ccl
