+++ ttnn/cpp/ttnn/operations/experimental/ccl/all_gather_async/all_gather_async.hpp
@@ -55,20 +55,6 @@ struct ExecuteAllGatherAsync {
         std::optional<tt::tt_metal::SubDeviceId> subdevice_id = std::nullopt);
 };
 
-struct ExecuteAllToAllAsync {
-    static ttnn::Tensor invoke(
-        const ttnn::Tensor& input_tensor,
-        ttnn::Tensor& persistent_intermediate_buffer,
-        ttnn::Tensor& persistent_output_buffer,
-        const int32_t in_dim,
-        const int32_t out_dim,
-        const GlobalSemaphore& multi_device_global_semaphore,
-        const uint32_t num_links = 1,
-        const std::optional<ttnn::MemoryConfig>& memory_config = std::nullopt,
-        const ttnn::ccl::Topology topology = ttnn::ccl::Topology::Ring,
-        std::optional<tt::tt_metal::SubDeviceId> subdevice_id = std::nullopt);
-};
-
 }  // namespace operations::experimental::ccl
 
 namespace experimental {
@@ -76,10 +62,6 @@ namespace experimental {
 constexpr auto all_gather_async = ttnn::register_operation<
     "ttnn::experimental::all_gather_async",
     ttnn::operations::experimental::ccl::ExecuteAllGatherAsync>();
-
-constexpr auto all_to_all_async = ttnn::register_operation<
-    "ttnn::experimental::all_to_all_async",
-    ttnn::operations::experimental::ccl::ExecuteAllToAllAsync>();
 
 }  // namespace experimental
 }  // namespace ttnn
