+++ ttnn/cpp/ttnn/operations/experimental/ccl/all_gather_async/device/all_gather_async_op.hpp
@@ -93,69 +93,6 @@ struct AllGatherAsync {
     AllGatherAsyncVersion select_version(const Tensor& input_tensor) const;
 };
 
-// New struct for AllToAllAsync
-struct AllToAllAsync {
-    std::vector<IDevice*> devices;
-    const uint32_t in_dim;
-    const uint32_t out_dim;
-    const uint32_t num_links;
-    const uint32_t ring_size;
-    const MemoryConfig output_mem_config;
-    const ccl::Topology topology;
-    const GlobalSemaphore semaphore;
-    std::optional<tt::tt_metal::SubDeviceId> sub_device_id;
-
-    AllToAllAsync(
-        std::vector<IDevice*> devices,
-        uint32_t in_dim,
-        uint32_t out_dim,
-        uint32_t num_links,
-        uint32_t ring_size,
-        MemoryConfig output_mem_config,
-        ccl::Topology topology,
-        GlobalSemaphore semaphore,
-        std::optional<tt::tt_metal::SubDeviceId>& sub_device_id) :
-        devices(std::move(devices)),
-        in_dim(in_dim),
-        out_dim(out_dim),
-        num_links(num_links),
-        ring_size(ring_size),
-        output_mem_config(output_mem_config),
-        topology(topology),
-        semaphore(semaphore),
-        sub_device_id(sub_device_id) {}
-
-    // Add attributes method for reflection
-    auto attributes() const {
-        using tt::stl::reflection::Attribute;
-        std::vector<std::tuple<std::string, Attribute>> attrs;
-
-        attrs.emplace_back("in_dim", in_dim);
-        attrs.emplace_back("out_dim", out_dim);
-        attrs.emplace_back("num_links", num_links);
-        attrs.emplace_back("ring_size", ring_size);
-        attrs.emplace_back("output_mem_config", output_mem_config);
-        attrs.emplace_back("topology", topology);
-        attrs.emplace_back("semaphore", semaphore);
-
-        return attrs;
-    }
-
-    // Method declarations (implementations will be needed elsewhere)
-    void validate_with_output_tensors(
-        const std::vector<Tensor>& input_tensors, const std::vector<std::optional<Tensor>>& output_tensors) const;
-    std::vector<ttnn::TensorSpec> compute_output_specs(const std::vector<Tensor>& input_tensors) const;
-    tt::tt_metal::operation::Hash compute_program_hash(const std::vector<Tensor>& input_tensors) const;
-    tt::tt_metal::operation::MeshWorkloadWithCallbacks create_mesh_workload(
-        const ttnn::MeshCoordinateRangeSet& tensor_coords,
-        const std::vector<Tensor>& input_tensors,
-        std::vector<Tensor>& output_tensors) const;
-    tt::tt_metal::operation::ProgramWithCallbacks create_program_at(
-        const ttnn::MeshCoordinate& coord,
-        const std::vector<Tensor>& input_tensors,
-        std::vector<Tensor>& output_tensors) const;
-};
-
 // All Gather Variants
 std::tuple<CoreRangeSet, std::vector<CoreCoord>> choose_worker_cores(
     size_t num_links,
@@ -203,23 +140,6 @@ tt::tt_metal::operation::ProgramWithCallbacks all_gather_async_llama_sharded(
     const GlobalSemaphore& semaphore,
     const std::optional<tt::tt_metal::SubDeviceId>& sub_device_id);
 
-// Add declaration for the AllToAll program function
-tt::tt_metal::operation::ProgramWithCallbacks all_to_all_async_minimal(
-    const Tensor& input_tensor,
-    Tensor& persistent_intermediate_buffer,
-    Tensor& persistent_output_buffer,
-    IDevice* target_device,
-    std::optional<IDevice*> forward_device,
-    std::optional<IDevice*> backward_device,
-    const uint32_t in_dim,
-    const uint32_t out_dim,
-    const uint32_t num_links,
-    const uint32_t ring_size,
-    const uint32_t ring_index,
-    ttnn::ccl::Topology topology,
-    const GlobalSemaphore& semaphore,
-    const std::optional<tt::tt_metal::SubDeviceId>& sub_device_id);
-
 namespace operations {
 namespace experimental {
 namespace ccl {
@@ -264,19 +184,6 @@ std::vector<Tensor> all_gather_async(
     const std::optional<ttnn::Tensor>& persistent_output_tensor = std::nullopt,
     const std::optional<MemoryConfig>& memory_config = std::nullopt,
     const std::optional<size_t> num_preferred_links = std::nullopt,
-    std::optional<tt::tt_metal::SubDeviceId> sub_device_id = std::nullopt);
-
-// Add declaration for all_to_all_async
-Tensor all_to_all_async(
-    const Tensor& input_tensor,
-    Tensor& persistent_intermediate_buffer,
-    Tensor& persistent_output_buffer,
-    const int32_t in_dim,
-    const int32_t out_dim,
-    const GlobalSemaphore& multi_device_global_semaphore,
-    const uint32_t num_links = 1,
-    const std::optional<MemoryConfig>& memory_config = std::nullopt,
-    const ttnn::ccl::Topology topology = ttnn::ccl::Topology::Ring,
     std::optional<tt::tt_metal::SubDeviceId> sub_device_id = std::nullopt);
 
 }  // namespace ccl
