+++ ttnn/cpp/ttnn/operations/experimental/ccl/all_to_all_async/all_to_all_async_pybind.hpp
@@ -1,0 +1,13 @@
+// SPDX-FileCopyrightText: © 2024 Tenstorrent AI ULC
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#pragma once
+
+#include "pybind11/pybind_fwd.hpp"
+
+namespace ttnn::operations::experimental::ccl {
+
+void py_bind_all_to_all_async(pybind11::module& module);
+
+}  // namespace ttnn::operations::experimental::ccl
