+++ ttnn/cpp/ttnn/operations/experimental/ccl/all_to_all_async/device/all_to_all_program_factory.cpp
@@ -9,7 +9,7 @@
 #include <tt-metalium/buffer.hpp>
 #include <tt-metalium/fabric.hpp>
 #include "ttnn/tensor/tensor_impl.hpp"
-#include "ttnn/operations/experimental/ccl/all_gather_async/device/all_gather_async_op.hpp"
+#include "ttnn/operations/experimental/ccl/all_to_all_async/device/all_to_all_async_op.hpp"
 #include "ttnn/operations/ccl/shared_with_host/hetergeneous_data_structs.hpp"
 #include "ttnn/operations/ccl/ccl_host_datastructures.hpp"
 #include "ttnn/operations/ccl/ccl_common.hpp"
@@ -279,7 +279,7 @@ tt::tt_metal::operation::ProgramWithCallbacks all_to_all_async_minimal(
     }
     auto worker_sender_reader_kernel_id = tt::tt_metal::CreateKernel(
         program,
-        "ttnn/cpp/ttnn/operations/experimental/ccl/all_gather_async/device/kernels/"
+        "ttnn/cpp/ttnn/operations/experimental/ccl/all_to_all_async/device/kernels/"
         "interleaved_all_to_all_reader.cpp",
         sender_worker_core_range,
         reader_kernel_config);
@@ -306,7 +306,7 @@ tt::tt_metal::operation::ProgramWithCallbacks all_to_all_async_minimal(
     }
     auto worker_sender_writer_kernel_id = tt::tt_metal::CreateKernel(
         program,
-        "ttnn/cpp/ttnn/operations/experimental/ccl/all_gather_async/device/kernels/"
+        "ttnn/cpp/ttnn/operations/experimental/ccl/all_to_all_async/device/kernels/"
         "interleaved_all_to_all_writer.cpp",
         sender_worker_core_range,
         writer_kernel_config);
@@ -325,7 +325,7 @@ tt::tt_metal::operation::ProgramWithCallbacks all_to_all_async_minimal(
 
     auto receiver_writer_kernel_id = tt::tt_metal::CreateKernel(
         program,
-        "ttnn/cpp/ttnn/operations/experimental/ccl/all_gather_async/device/kernels/"
+        "ttnn/cpp/ttnn/operations/experimental/ccl/all_to_all_async/device/kernels/"
         "interleaved_all_to_all_receiver_writer.cpp",
         receiver_worker_core_range,
         receiver_writer_kernel_config);
@@ -343,7 +343,7 @@ tt::tt_metal::operation::ProgramWithCallbacks all_to_all_async_minimal(
         pages_per_packet};
     auto receiver_reader_kernel_id = tt::tt_metal::CreateKernel(
         program,
-        "ttnn/cpp/ttnn/operations/experimental/ccl/all_gather_async/device/kernels/"
+        "ttnn/cpp/ttnn/operations/experimental/ccl/all_to_all_async/device/kernels/"
         "interleaved_all_to_all_receiver_reader.cpp",
         receiver_worker_core_range,
         receiver_reader_kernel_config);
