+++ ttnn/cpp/ttnn/operations/experimental/ccl/all_to_all_async/device/kernels/interleaved_all_to_all_writer.cpp
@@ -6,10 +6,8 @@
 #include <tt-metalium/buffer_types.hpp>
 #include "tt_metal/fabric/hw/inc/edm_fabric/fabric_connection_manager.hpp"
 #include "cpp/ttnn/operations/ccl/common/interpreter_backends/kernel_common/noc_addr.hpp"
-#include "minimal_ccl_common.hpp"
 #include <cstdint>
 #include <utility>
-#include "tt_metal/tools/profiler/kernel_profiler.hpp"
 
 using address_t = uint32_t;
 using tt::tt_metal::BufferType;
@@ -149,7 +147,6 @@ void kernel_main() {
     uint32_t backward_hops = 1;
     uint32_t dst_ring_id;
     for (uint32_t i = 0; i < ring_size - 1; ++i) {
-        DeviceZoneScopedN("WriteToDevice");
         if (forward_hops == num_targets_forward_direction + 1) {
             cur_is_forward = false;
         }
